Feature/check for nil (#11)
* Initial implementation of Raven compiler

* feat: release v0.2.0 - core features and stability improvements

Integrated new language features, optimized the middle-end, and improved
compiler stability.

Features:
- Added 'while' loops, string indexing (str[i]), and 'sizeof' builtin.
- Added syscall support with improved handling.
- Added Intel syntax support and new CLI flags (e.g., -o).

Internal & Optimization:
- Implemented middle-end optimization and dead code elimination.
- Refactored core modules for safer memory allocation.
- Standardized documentation using Doxygen style.

Fixes:
- Resolved compiler segfaults and improved memory safety.
- Fixed assembly indentation and lexer whitespace issues.
- Integrated comprehensive unit tests.

* fix: final clean up of merge markers

* ver: 0.3.0
Added:
- optional type hints
- null type

* docs: create CONTRIBUTING.md (#7)

* feat: rename "null" to "nil" (#8)

* feat: unstable check-for-nil

---------

+++ src/raven/token.c
@@ -14,7 +14,7 @@ const char *const token_strings[] = {TOKEN_LIST};
 static const int token_count = sizeof(token_strings) / sizeof(token_strings[0]);
 
 const char *token_type_to_string(csq_tktype type) {
-  if (type >= 0 && type < token_count) {
+  if (type >= 0 && (const int)type < token_count) {
     return token_strings[type];
   }
   return "UNKNOWN";
